requirements.txt
Django~=5.1.3
djangorestframework~=3.15.2
djangorestframework-simplejwt-5.3.1
pillow~=11.0.0
psycopg2-binary
python-dotenv~=1.0.1
pip~=24.3.1
Jinja2~=3.1.4
filelock~=3.14.0
redis~=5.1.1
asgiref~=3.8.1
sqlparse~=0.5.1
psycopg2~=2.9.9
colorama~=0.4.6
PyYAML~=6.0.2
requests~=2.31.0
click~=8.1.7
pytz~=2024.2
django-filter~=24.3